Fixed issues: - not filling the proposal from the start prevent the thumbnail url to be filled. Thus, we have to load all proposals at start - disabling sync button at start (on the return) blocks the syncClick function binding. Thus, had to displace the disable on the useEffect
