--- v0_1/assets/main-75dafeac.js
@@ -1,229 +1,0 @@
-var or=Object.defineProperty;var ir=(e,t,r)=>t in e?or(e,t,{enumerable:!0,configurable:!0,writable:!0,value:r}):e[t]=r;var te=(e,t,r)=>(ir(e,typeof t!="symbol"?t+"":t,r),r);import{R as ot,r as G,j as h,F as pe,a as Ue,b as ie,c as ar}from"./index-340c65dc.js";/**
- * @remix-run/router v1.3.2
- *
- * Copyright (c) Remix Software Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE.md file in the root directory of this source tree.
- *
- * @license MIT
- */function $e(){return $e=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var r=arguments[t];for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&(e[n]=r[n])}return e},$e.apply(this,arguments)}var ge;(function(e){e.Pop="POP",e.Push="PUSH",e.Replace="REPLACE"})(ge||(ge={}));const xt="popstate";function sr(e){e===void 0&&(e={});function t(o,i){let{pathname:a="/",search:c="",hash:l=""}=pt(o.location.hash.substr(1));return it("",{pathname:a,search:c,hash:l},i.state&&i.state.usr||null,i.state&&i.state.key||"default")}function r(o,i){let a=o.document.querySelector("base"),c="";if(a&&a.getAttribute("href")){let l=o.location.href,d=l.indexOf("#");c=d===-1?l:l.slice(0,d)}return c+"#"+(typeof i=="string"?i:Vt(i))}function n(o,i){cr(o.pathname.charAt(0)==="/","relative pathnames are not supported in hash history.push("+JSON.stringify(i)+")")}return ur(t,r,n,e)}function Bt(e,t){if(e===!1||e===null||typeof e>"u")throw new Error(t)}function cr(e,t){if(!e){typeof console<"u"&&console.warn(t);try{throw new Error(t)}catch{}}}function lr(){return Math.random().toString(36).substr(2,8)}function At(e,t){return{usr:e.state,key:e.key,idx:t}}function it(e,t,r,n){return r===void 0&&(r=null),$e({pathname:typeof e=="string"?e:e.pathname,search:"",hash:""},typeof t=="string"?pt(t):t,{state:r,key:t&&t.key||n||lr()})}function Vt(e){let{pathname:t="/",search:r="",hash:n=""}=e;return r&&r!=="?"&&(t+=r.charAt(0)==="?"?r:"?"+r),n&&n!=="#"&&(t+=n.charAt(0)==="#"?n:"#"+n),t}function pt(e){let t={};if(e){let r=e.indexOf("#");r>=0&&(t.hash=e.substr(r),e=e.substr(0,r));let n=e.indexOf("?");n>=0&&(t.search=e.substr(n),e=e.substr(0,n)),e&&(t.pathname=e)}return t}function ur(e,t,r,n){n===void 0&&(n={});let{window:o=document.defaultView,v5Compat:i=!1}=n,a=o.history,c=ge.Pop,l=null,d=m();d==null&&(d=0,a.replaceState($e({},a.state,{idx:d}),""));function m(){return(a.state||{idx:null}).idx}function S(){c=ge.Pop;let p=m(),A=p==null?null:p-d;d=p,l&&l({action:c,location:b.location,delta:A})}function P(p,A){c=ge.Push;let w=it(b.location,p,A);r&&r(w,p),d=m()+1;let z=At(w,d),F=b.createHref(w);try{a.pushState(z,"",F)}catch{o.location.assign(F)}i&&l&&l({action:c,location:b.location,delta:1})}function D(p,A){c=ge.Replace;let w=it(b.location,p,A);r&&r(w,p),d=m();let z=At(w,d),F=b.createHref(w);a.replaceState(z,"",F),i&&l&&l({action:c,location:b.location,delta:0})}function E(p){let A=o.location.origin!=="null"?o.location.origin:o.location.href,w=typeof p=="string"?p:Vt(p);return Bt(A,"No window.location.(origin|href) available to create URL for href: "+w),new URL(w,A)}let b={get action(){return c},get location(){return e(o,a)},listen(p){if(l)throw new Error("A history only accepts one active listener");return o.addEventListener(xt,S),l=p,()=>{o.removeEventListener(xt,S),l=null}},createHref(p){return t(o,p)},createURL:E,encodeLocation(p){let A=E(p);return{pathname:A.pathname,search:A.search,hash:A.hash}},push:P,replace:D,go(p){return a.go(p)}};return b}var It;(function(e){e.data="data",e.deferred="deferred",e.redirect="redirect",e.error="error"})(It||(It={}));function dr(e,t){if(t==="/")return e;if(!e.toLowerCase().startsWith(t.toLowerCase()))return null;let r=t.endsWith("/")?t.length-1:t.length,n=e.charAt(r);return n&&n!=="/"?null:e.slice(r)||"/"}const pr=["post","put","patch","delete"];[...pr];/**
- * React Router v6.8.1
- *
- * Copyright (c) Remix Software Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE.md file in the root directory of this source tree.
- *
- * @license MIT
- */function fr(e,t){return e===t&&(e!==0||1/e===1/t)||e!==e&&t!==t}const hr=typeof Object.is=="function"?Object.is:fr,{useState:mr,useEffect:gr,useLayoutEffect:yr,useDebugValue:vr}=ot;function wr(e,t,r){const n=t(),[{inst:o},i]=mr({inst:{value:n,getSnapshot:t}});return yr(()=>{o.value=n,o.getSnapshot=t,Qe(o)&&i({inst:o})},[e,n,t]),gr(()=>(Qe(o)&&i({inst:o}),e(()=>{Qe(o)&&i({inst:o})})),[e]),vr(n),n}function Qe(e){const t=e.getSnapshot,r=e.value;try{const n=t();return!hr(r,n)}catch{return!0}}function br(e,t,r){return t()}const Sr=typeof window<"u"&&typeof window.document<"u"&&typeof window.document.createElement<"u",xr=!Sr,Ar=xr?br:wr;"useSyncExternalStore"in ot&&(e=>e.useSyncExternalStore)(ot);const Ir=G.createContext(null),Ht=G.createContext(null);function Cr(){return G.useContext(Ht)!=null}var Ct;(function(e){e.UseBlocker="useBlocker",e.UseRevalidator="useRevalidator"})(Ct||(Ct={}));var kt;(function(e){e.UseLoaderData="useLoaderData",e.UseActionData="useActionData",e.UseRouteError="useRouteError",e.UseNavigation="useNavigation",e.UseRouteLoaderData="useRouteLoaderData",e.UseMatches="useMatches",e.UseRevalidator="useRevalidator"})(kt||(kt={}));function kr(e){let{basename:t="/",children:r=null,location:n,navigationType:o=ge.Pop,navigator:i,static:a=!1}=e;Cr()&&Bt(!1);let c=t.replace(/^\/*/,"/"),l=G.useMemo(()=>({basename:c,navigator:i,static:a}),[c,i,a]);typeof n=="string"&&(n=pt(n));let{pathname:d="/",search:m="",hash:S="",state:P=null,key:D="default"}=n,E=G.useMemo(()=>{let b=dr(d,c);return b==null?null:{pathname:b,search:m,hash:S,state:P,key:D}},[c,d,m,S,P,D]);return E==null?null:G.createElement(Ir.Provider,{value:l},G.createElement(Ht.Provider,{children:r,value:{location:E,navigationType:o}}))}var Pt;(function(e){e[e.pending=0]="pending",e[e.success=1]="success",e[e.error=2]="error"})(Pt||(Pt={}));new Promise(()=>{});/**
- * React Router DOM v6.8.1
- *
- * Copyright (c) Remix Software Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE.md file in the root directory of this source tree.
- *
- * @license MIT
- */function Pr(e){let{basename:t,children:r,window:n}=e,o=G.useRef();o.current==null&&(o.current=sr({window:n,v5Compat:!0}));let i=o.current,[a,c]=G.useState({action:i.action,location:i.location});return G.useLayoutEffect(()=>i.listen(c),[i]),G.createElement(kr,{basename:t,children:r,location:a.location,navigationType:a.action,navigator:i})}var Et;(function(e){e.UseScrollRestoration="useScrollRestoration",e.UseSubmitImpl="useSubmitImpl",e.UseFetcher="useFetcher"})(Et||(Et={}));var Rt;(function(e){e.UseFetchers="useFetchers",e.UseScrollRestoration="useScrollRestoration"})(Rt||(Rt={}));var Wt={};(function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.ReferenceType=void 0,function(t){t.INLINE="Inline",t.REMOTE="Remote"}(e.ReferenceType||(e.ReferenceType={}))})(Wt);var Te,Er=new Uint8Array(16);function Rr(){if(!Te&&(Te=typeof crypto<"u"&&crypto.getRandomValues&&crypto.getRandomValues.bind(crypto)||typeof msCrypto<"u"&&typeof msCrypto.getRandomValues=="function"&&msCrypto.getRandomValues.bind(msCrypto),!Te))throw new Error("crypto.getRandomValues() not supported. See https://github.com/uuidjs/uuid#getrandomvalues-not-supported");return Te(Er)}const Or=/^(?:[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|00000000-0000-0000-0000-000000000000)$/i;function Tr(e){return typeof e=="string"&&Or.test(e)}var K=[];for(var et=0;et<256;++et)K.push((et+256).toString(16).substr(1));function Fr(e){var t=arguments.length>1&&arguments[1]!==void 0?arguments[1]:0,r=(K[e[t+0]]+K[e[t+1]]+K[e[t+2]]+K[e[t+3]]+"-"+K[e[t+4]]+K[e[t+5]]+"-"+K[e[t+6]]+K[e[t+7]]+"-"+K[e[t+8]]+K[e[t+9]]+"-"+K[e[t+10]]+K[e[t+11]]+K[e[t+12]]+K[e[t+13]]+K[e[t+14]]+K[e[t+15]]).toLowerCase();if(!Tr(r))throw TypeError("Stringified UUID is invalid");return r}function at(e,t,r){e=e||{};var n=e.random||(e.rng||Rr)();if(n[6]=n[6]&15|64,n[8]=n[8]&63|128,t){r=r||0;for(var o=0;o<16;++o)t[r+o]=n[o];return t}return Fr(n)}const _r=e=>[e[0],e[4],e[8],e[12],e[1],e[5],e[9],e[13],e[2],e[6],e[10],e[14],e[3],e[7],e[11],e[15]],Lr=e=>{const t={},r=[],n=[];for(const o of e.indices){const[i,a,c]=e.vertices.slice(o*3,o*3+3);t[i]||(t[i]={}),t[i][a]||(t[i][a]={}),t[i][a][c]||(t[i][a][c]=n.length/3,n.push(i,a,c)),r.push(t[i][a][c])}return{...e,indices:r,vertices:n}},jr=e=>{let t=[],r=[];for(const n of e){const o=r.length/3;t=t.concat(n.indices.map(i=>i+o)),r=r.concat(n.vertices)}return{...e[0],indices:t,vertices:r}},Nr=async e=>{const t={id:at(),properties:{category:"ConceptualRoot"},children:[]},r={[t.id]:t},n={rootElement:t.id,elements:r};let o=await WSM.Geom.Transf3d();for(const i of e){const a=Lr(jr(i.meshes)),c={id:at(),children:[],properties:{category:"ConceptualElement",geometry:{type:Wt.ReferenceType.INLINE,format:"Mesh",verts:a.vertices,faces:a.indices}}};i.transforms[0].data&&(o.data=i.transforms[0].data),r[c.id]={...c},t.children.push({id:c.id,transform:_r(o.data)})}return n},$r=e=>{const t=e.split(":");if(t.length===6){const[,r,n,o,i,a]=t;return{namespace:r,system:n,authcontext:o,id:i,revision:a}}else if(t.length===5){const[,r,n,o,i]=t;return{namespace:r,system:n,id:o,revision:i}}throw new Error("urn is not of length 5 or 6")},tt=0,Fe=.3047999995367042,re={FORMA_CONTEXT:"Forma Context",FORMA_TERRAIN:"Forma Terrain",SURROUNDING_BUILDINGS:"Surrounding Buildings",FORMA_PROPOSAL_BUILDINGS:"Forma Proposal Buildings",FORMA_AUTO_BUILDINGS:"Forma Auto Buildings",FORMA_BUILDINGS:"Forma Buildings",FORMA_SITE_LIMIT:"site_limit",FORMA_BUILDING:"building",FORMA_VEGETATION:"vegetation",FORMA_GENERIC:"generic",FORMA_ROAD:"road",FORMA_RAILS:"rails",FORMA_PROPERTY_BOUNDARY:"property_boundary",FORMA_ZONE:"zone",FORMA_BUILDING_ENVELOPE:"building_envelope"};class Mr{getCookie(t){const r=t.length+1;return document.cookie.split(";").map(n=>n.trim()).filter(n=>n.substring(0,r)===`${t}=`).map(n=>decodeURIComponent(n.substring(r)))[0]||null}accessSpacemaker(t){let r=null;if(t){r=window.open("https://app.spacemaker.ai/auth/login?rd=https%3A%2F%2Fapp.spacemaker.ai%2Fprojects","_blank","width= 500px, height=500px");let n=setInterval(()=>{this.getCookie("ajs_user_id")!==null&&(clearInterval(n),r!==null&&r.close())},1e3)}else{const o="https://local.spacemaker.ai:3001?loggedIn=1";window.location.replace(`https://app.spacemaker.ai/auth/login?rd=${o}`)}}async save({projectId:t,proposalId:r},n){FormItInterface.CallMethod("FormitPlugin.GetAllGeometryInformation",[],async()=>{this.getFormitGeometry(re.FORMA_BUILDINGS,(o,i)=>{i||(i={});let a=0;this.createIntegrateAPIElementAndUpdateProposal({formitGeometry:o,proposalId:r,projectId:t,polygonData:i,objectId:a,callback:n})})})}async getFormitGeometry(t,r){this.hideLayersBeforeSave().then(async n=>{FormIt.Layers.SetLayerVisibility(t,!1).then(()=>{FormItInterface.CallMethod("FormitPlugin.GetAllGeometryInformation",[],async o=>{let i=JSON.parse(o);FormIt.Layers.SetLayerVisibility(t,!0).then(()=>{this.returnLayersToPreviousVibility(n).then(()=>{this.getPolygonData(tt,i,r)})})})})})}async getAllGeometryInformations(){var t=await WSM.Utils.GetAllGeometryInformation(tt);return t===null&&(t=[]),t}async returnLayersToPreviousVibility(t){await t.forEach(async r=>{r.layerData&&(r.layerData.Visible=r.previousVisiblity,await FormIt.Layers.SetLayersVisibility([r.layerData]))})}async hideLayersBeforeSave(){const t=[re.FORMA_CONTEXT,re.FORMA_TERRAIN,re.SURROUNDING_BUILDINGS,re.FORMA_AUTO_BUILDINGS,re.FORMA_PROPOSAL_BUILDINGS,re.FORMA_SITE_LIMIT,re.FORMA_BUILDING,re.FORMA_VEGETATION,re.FORMA_GENERIC,re.FORMA_ROAD,re.FORMA_RAILS,re.FORMA_PROPERTY_BOUNDARY,re.FORMA_ZONE,re.FORMA_BUILDING_ENVELOPE];let r=[];for(const n of t){let o=await this.mapLayerVisibility(n);o.layerData&&r.push(o)}return r}async mapLayerVisibility(t){const r=await FormIt.Layers.GetLayerID(t);let n=!1,o;return r!=WSM.INVALID_ID&&(o=await FormIt.Layers.GetLayerData(r),n=o.Visible,o.Visible=!1,await FormIt.Layers.SetLayersVisibility([o])),{layerData:o,previousVisiblity:n}}async getPolygonData(t=WSM.INVALID_ID,r,n){FormItInterface.CallMethod("FormitPlugin.ComputeGeometryFromLevels",[t],o=>{let i={},a=JSON.parse(o);a&&a.map(c=>{const{outer_loop:l,inner_loops:d}=c.second[0],S=[l.vertices.map(P=>[P.x*Fe,P.y*Fe])];if(d.length>0){const P=d[0].vertices.map(D=>[D.x*Fe,D.y*Fe]);S.push(P)}return S}),n&&n(r,i)})}async saveTemp(t,r){await FormIt.GroupEdit.EndEditInContext(),await FormIt.Selection.SelectAll();var n=await WSM.APISaveToStringReadOnly(r,[t]);return await FormIt.Selection.ClearSelections(),n}async generateStorageReference(t,r,n){const o=new Blob([t],{type:"application/octet-stream"}),i=new File([o],"internalRepresentation.wsm");await this.createS3Object(i,r).then(a=>{n&&n(a)})}async createS3Object(t,r){const n=JSON.stringify({projectId:r});try{const o=await fetch("/api/spacemaker-object-storage/v1/",{method:"POST",body:n,headers:{"Content-Type":"application/json","Set-Cookie":"true"}});if(o.ok){const{id:i,url:a,fields:c}=await o.json(),l=new FormData;if(Object.keys(c).forEach(m=>{l.append(m,c[m])}),l.append("file",t),(await fetch(a,{method:"POST",body:l,headers:{"Set-Cookie":"true"}})).ok)return i}}catch{return!1}return!1}generatePayload(t,r){Nr(t).then(n=>{r&&r(n)})}getWSMLayerID(t,r){const n=WSM.APIGetAllObjectsByTypeReadOnly(t,WSM.nLayerType);let o=WSM.INVALID_ID;for(let i=0;i<n.length;i++)if(WSM.APIGetLayerDataReadOnly(t,n[i]).name==r){o=n[i];break}return o}async createIntegrateAPIElementAndUpdateProposal({formitGeometry:t,proposalId:r,projectId:n,polygonData:o,objectId:i,callback:a}){this.saveTemp(i,tt).then(c=>{if(!c){console.error("Can't save temporary wsm file.");return}this.generatePayload(t,async l=>{this.generateStorageReference(c,n,async d=>{if(!d){a&&a(!1);return}this.createOrUpdateElement(n,l,d,o).then(async m=>{m?this.updateProposalElement({elementId:r,authContext:n,createdUrn:m.urn}).then(S=>{a&&a(S)}):a&&a(!1)})})})})}async createOrUpdateElement(t,r,n,o){let i=await r;const a=i.elements[i.rootElement].properties;a.spacemakerObjectStorageReferences=[n],a.spacemakerObjectStorageReferenceFormats=["wsm"],o.length>0&&(a.areaStatsReps={grossFloorPolygons:o});const c=await this.getUploadLink(t);if(c){const l=await this.uploadData(c.url,JSON.stringify(i));try{if(l){const d=`/api/integrate/elements?version=2&authcontext=${t}&s3Id=${c.id}`,m=await fetch(d,{method:"POST"});if(m.ok)return await m.json()}}catch{return!1}}return!1}async updateProposalElement({elementId:t,authContext:r,createdUrn:n}){var i;const o=await this.getProposalElement(t,r);if(o===null){console.error("proposal can't be retrieved to be updated");return}for(const[a,c]of Object.entries(o))if(((i=c.properties)==null?void 0:i.category)==="proposal"){const{revision:l}=$r(a);if(n)c.children.push({urn:n,key:at()});else return console.error("not a valid update, bad parameters"),!1;try{return(await fetch(`/api/proposal/elements/${t}/revisions/${l}?version=2&authcontext=${r}`,{method:"PUT",body:JSON.stringify(c)})).ok}catch{return!1}}}async getUploadLink(t){try{const r=await fetch(`/api/integrate/upload_link?authcontext=${t}`,{method:"GET",headers:{"Content-Type":"application/json","Set-Cookie":"true"}});if(r.ok)return await r.json()}catch{return!1}return!1}async uploadData(t,r){try{return(await fetch(t,{method:"PUT",body:r})).ok}catch{return!1}}async getProposalElement(t,r){try{const n=`/api/proposal/elements/${t}?authcontext=${r}`,o=await fetch(n);return o.ok?await o.json():null}catch{return null}}}const ft=new Mr;class Dr extends G.Component{constructor(t){super(t),this.state={loggedIn:!1}}static login(){return ft.accessSpacemaker(!1),null}render(){return h(pe,{})}}class ht{constructor(){te(this,"id");te(this,"urn");te(this,"name");te(this,"metadata");te(this,"version");this.id="",this.urn="",this.name="",this.version="1",this.metadata=null}Fill(t,r,n,o,i){if(this.id=t,this.version=n,r===null||r===""){let a=t.split("prop_");this.name=`prop_${a[1].substring(0,2)}`}else this.name=r;this.urn=i,this.metadata=o}}const _e="https://local.spacemaker.ai:3001",Ur="https://app.spacemaker.ai";class zr{getWorkspaces(){return fetch(`${_e}/api/workspaces`).then(t=>{if(!t.ok)throw new Error(t.statusText);return t.json()})}getProjects(t){return fetch(`${_e}/api/projects?customer=${t}&include_archived=false`).then(r=>{if(!r.ok)throw new Error(r.statusText);return r.json()})}countProposals(t){return fetch(`${_e}/api/proposal/elements/count?authcontext=${t}`).then(r=>{if(!r.ok)throw new Error(r.statusText);return r.json()})}getProposals(t){return fetch(`${_e}/api/proposal/elements?authcontext=${t}&version=2`).then(r=>{if(!r.ok)throw new Error(r.statusText);return r.json()})}FormatThumbnailUrl(t,r){return`${Ur}/api/thumbnails/v2/${r}?size=170&authcontext=${t}&projectId=${t}`}}const ve=new zr;var Pe={},Gr={get exports(){return Pe},set exports(e){Pe=e}},$={};/** @license React v16.13.1
- * react-is.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var W=typeof Symbol=="function"&&Symbol.for,mt=W?Symbol.for("react.element"):60103,gt=W?Symbol.for("react.portal"):60106,ze=W?Symbol.for("react.fragment"):60107,Ge=W?Symbol.for("react.strict_mode"):60108,Be=W?Symbol.for("react.profiler"):60114,Ve=W?Symbol.for("react.provider"):60109,He=W?Symbol.for("react.context"):60110,yt=W?Symbol.for("react.async_mode"):60111,We=W?Symbol.for("react.concurrent_mode"):60111,Ye=W?Symbol.for("react.forward_ref"):60112,Xe=W?Symbol.for("react.suspense"):60113,Br=W?Symbol.for("react.suspense_list"):60120,Ze=W?Symbol.for("react.memo"):60115,Je=W?Symbol.for("react.lazy"):60116,Vr=W?Symbol.for("react.block"):60121,Hr=W?Symbol.for("react.fundamental"):60117,Wr=W?Symbol.for("react.responder"):60118,Yr=W?Symbol.for("react.scope"):60119;function ae(e){if(typeof e=="object"&&e!==null){var t=e.$$typeof;switch(t){case mt:switch(e=e.type,e){case yt:case We:case ze:case Be:case Ge:case Xe:return e;default:switch(e=e&&e.$$typeof,e){case He:case Ye:case Je:case Ze:case Ve:return e;default:return t}}case gt:return t}}}function Yt(e){return ae(e)===We}$.AsyncMode=yt;$.ConcurrentMode=We;$.ContextConsumer=He;$.ContextProvider=Ve;$.Element=mt;$.ForwardRef=Ye;$.Fragment=ze;$.Lazy=Je;$.Memo=Ze;$.Portal=gt;$.Profiler=Be;$.StrictMode=Ge;$.Suspense=Xe;$.isAsyncMode=function(e){return Yt(e)||ae(e)===yt};$.isConcurrentMode=Yt;$.isContextConsumer=function(e){return ae(e)===He};$.isContextProvider=function(e){return ae(e)===Ve};$.isElement=function(e){return typeof e=="object"&&e!==null&&e.$$typeof===mt};$.isForwardRef=function(e){return ae(e)===Ye};$.isFragment=function(e){return ae(e)===ze};$.isLazy=function(e){return ae(e)===Je};$.isMemo=function(e){return ae(e)===Ze};$.isPortal=function(e){return ae(e)===gt};$.isProfiler=function(e){return ae(e)===Be};$.isStrictMode=function(e){return ae(e)===Ge};$.isSuspense=function(e){return ae(e)===Xe};$.isValidElementType=function(e){return typeof e=="string"||typeof e=="function"||e===ze||e===We||e===Be||e===Ge||e===Xe||e===Br||typeof e=="object"&&e!==null&&(e.$$typeof===Je||e.$$typeof===Ze||e.$$typeof===Ve||e.$$typeof===He||e.$$typeof===Ye||e.$$typeof===Hr||e.$$typeof===Wr||e.$$typeof===Yr||e.$$typeof===Vr)};$.typeOf=ae;(function(e){e.exports=$})(Gr);function Xr(e){function t(y,f,g,C,s){for(var _=0,u=0,B=0,L=0,N,k,X=0,oe=0,T,q=T=N=0,j=0,Z=0,Ie=0,J=0,Oe=g.length,Ce=Oe-1,le,x="",V="",Ke="",qe="",fe;j<Oe;){if(k=g.charCodeAt(j),j===Ce&&u+L+B+_!==0&&(u!==0&&(k=u===47?10:47),L=B=_=0,Oe++,Ce++),u+L+B+_===0){if(j===Ce&&(0<Z&&(x=x.replace(P,"")),0<x.trim().length)){switch(k){case 32:case 9:case 59:case 13:case 10:break;default:x+=g.charAt(j)}k=59}switch(k){case 123:for(x=x.trim(),N=x.charCodeAt(0),T=1,J=++j;j<Oe;){switch(k=g.charCodeAt(j)){case 123:T++;break;case 125:T--;break;case 47:switch(k=g.charCodeAt(j+1)){case 42:case 47:e:{for(q=j+1;q<Ce;++q)switch(g.charCodeAt(q)){case 47:if(k===42&&g.charCodeAt(q-1)===42&&j+2!==q){j=q+1;break e}break;case 10:if(k===47){j=q+1;break e}}j=q}}break;case 91:k++;case 40:k++;case 34:case 39:for(;j++<Ce&&g.charCodeAt(j)!==k;);}if(T===0)break;j++}switch(T=g.substring(J,j),N===0&&(N=(x=x.replace(S,"").trim()).charCodeAt(0)),N){case 64:switch(0<Z&&(x=x.replace(P,"")),k=x.charCodeAt(1),k){case 100:case 109:case 115:case 45:Z=f;break;default:Z=me}if(T=t(f,Z,T,k,s+1),J=T.length,0<se&&(Z=r(me,x,Ie),fe=c(3,T,Z,f,Q,H,J,k,s,C),x=Z.join(""),fe!==void 0&&(J=(T=fe.trim()).length)===0&&(k=0,T="")),0<J)switch(k){case 115:x=x.replace(O,a);case 100:case 109:case 45:T=x+"{"+T+"}";break;case 107:x=x.replace(w,"$1 $2"),T=x+"{"+T+"}",T=Y===1||Y===2&&i("@"+T,3)?"@-webkit-"+T+"@"+T:"@"+T;break;default:T=x+T,C===112&&(T=(V+=T,""))}else T="";break;default:T=t(f,r(f,x,Ie),T,C,s+1)}Ke+=T,T=Ie=Z=q=N=0,x="",k=g.charCodeAt(++j);break;case 125:case 59:if(x=(0<Z?x.replace(P,""):x).trim(),1<(J=x.length))switch(q===0&&(N=x.charCodeAt(0),N===45||96<N&&123>N)&&(J=(x=x.replace(" ",":")).length),0<se&&(fe=c(1,x,f,y,Q,H,V.length,C,s,C))!==void 0&&(J=(x=fe.trim()).length)===0&&(x="\0\0"),N=x.charCodeAt(0),k=x.charCodeAt(1),N){case 0:break;case 64:if(k===105||k===99){qe+=x+g.charAt(j);break}default:x.charCodeAt(J-1)!==58&&(V+=o(x,N,k,x.charCodeAt(2)))}Ie=Z=q=N=0,x="",k=g.charCodeAt(++j)}}switch(k){case 13:case 10:u===47?u=0:1+N===0&&C!==107&&0<x.length&&(Z=1,x+="\0"),0<se*ye&&c(0,x,f,y,Q,H,V.length,C,s,C),H=1,Q++;break;case 59:case 125:if(u+L+B+_===0){H++;break}default:switch(H++,le=g.charAt(j),k){case 9:case 32:if(L+_+u===0)switch(X){case 44:case 58:case 9:case 32:le="";break;default:k!==32&&(le=" ")}break;case 0:le="\\0";break;case 12:le="\\f";break;case 11:le="\\v";break;case 38:L+u+_===0&&(Z=Ie=1,le="\f"+le);break;case 108:if(L+u+_+ee===0&&0<q)switch(j-q){case 2:X===112&&g.charCodeAt(j-3)===58&&(ee=X);case 8:oe===111&&(ee=oe)}break;case 58:L+u+_===0&&(q=j);break;case 44:u+B+L+_===0&&(Z=1,le+="\r");break;case 34:case 39:u===0&&(L=L===k?0:L===0?k:L);break;case 91:L+u+B===0&&_++;break;case 93:L+u+B===0&&_--;break;case 41:L+u+_===0&&B--;break;case 40:if(L+u+_===0){if(N===0)switch(2*X+3*oe){case 533:break;default:N=1}B++}break;case 64:u+B+L+_+q+T===0&&(T=1);break;case 42:case 47:if(!(0<L+_+B))switch(u){case 0:switch(2*k+3*g.charCodeAt(j+1)){case 235:u=47;break;case 220:J=j,u=42}break;case 42:k===47&&X===42&&J+2!==j&&(g.charCodeAt(J+2)===33&&(V+=g.substring(J,j+1)),le="",u=0)}}u===0&&(x+=le)}oe=X,X=k,j++}if(J=V.length,0<J){if(Z=f,0<se&&(fe=c(2,V,Z,y,Q,H,J,C,s,C),fe!==void 0&&(V=fe).length===0))return qe+V+Ke;if(V=Z.join(",")+"{"+V+"}",Y*ee!==0){switch(Y!==2||i(V,2)||(ee=0),ee){case 111:V=V.replace(F,":-moz-$1")+V;break;case 112:V=V.replace(z,"::-webkit-input-$1")+V.replace(z,"::-moz-$1")+V.replace(z,":-ms-input-$1")+V}ee=0}}return qe+V+Ke}function r(y,f,g){var C=f.trim().split(p);f=C;var s=C.length,_=y.length;switch(_){case 0:case 1:var u=0;for(y=_===0?"":y[0]+" ";u<s;++u)f[u]=n(y,f[u],g).trim();break;default:var B=u=0;for(f=[];u<s;++u)for(var L=0;L<_;++L)f[B++]=n(y[L]+" ",C[u],g).trim()}return f}function n(y,f,g){var C=f.charCodeAt(0);switch(33>C&&(C=(f=f.trim()).charCodeAt(0)),C){case 38:return f.replace(A,"$1"+y.trim());case 58:return y.trim()+f.replace(A,"$1"+y.trim());default:if(0<1*g&&0<f.indexOf("\f"))return f.replace(A,(y.charCodeAt(0)===58?"":"$1")+y.trim())}return y+f}function o(y,f,g,C){var s=y+";",_=2*f+3*g+4*C;if(_===944){y=s.indexOf(":",9)+1;var u=s.substring(y,s.length-1).trim();return u=s.substring(0,y).trim()+u+";",Y===1||Y===2&&i(u,1)?"-webkit-"+u+u:u}if(Y===0||Y===2&&!i(s,1))return s;switch(_){case 1015:return s.charCodeAt(10)===97?"-webkit-"+s+s:s;case 951:return s.charCodeAt(3)===116?"-webkit-"+s+s:s;case 963:return s.charCodeAt(5)===110?"-webkit-"+s+s:s;case 1009:if(s.charCodeAt(4)!==100)break;case 969:case 942:return"-webkit-"+s+s;case 978:return"-webkit-"+s+"-moz-"+s+s;case 1019:case 983:return"-webkit-"+s+"-moz-"+s+"-ms-"+s+s;case 883:if(s.charCodeAt(8)===45)return"-webkit-"+s+s;if(0<s.indexOf("image-set(",11))return s.replace(ue,"$1-webkit-$2")+s;break;case 932:if(s.charCodeAt(4)===45)switch(s.charCodeAt(5)){case 103:return"-webkit-box-"+s.replace("-grow","")+"-webkit-"+s+"-ms-"+s.replace("grow","positive")+s;case 115:return"-webkit-"+s+"-ms-"+s.replace("shrink","negative")+s;case 98:return"-webkit-"+s+"-ms-"+s.replace("basis","preferred-size")+s}return"-webkit-"+s+"-ms-"+s+s;case 964:return"-webkit-"+s+"-ms-flex-"+s+s;case 1023:if(s.charCodeAt(8)!==99)break;return u=s.substring(s.indexOf(":",15)).replace("flex-","").replace("space-between","justify"),"-webkit-box-pack"+u+"-webkit-"+s+"-ms-flex-pack"+u+s;case 1005:return E.test(s)?s.replace(D,":-webkit-")+s.replace(D,":-moz-")+s:s;case 1e3:switch(u=s.substring(13).trim(),f=u.indexOf("-")+1,u.charCodeAt(0)+u.charCodeAt(f)){case 226:u=s.replace(v,"tb");break;case 232:u=s.replace(v,"tb-rl");break;case 220:u=s.replace(v,"lr");break;default:return s}return"-webkit-"+s+"-ms-"+u+s;case 1017:if(s.indexOf("sticky",9)===-1)break;case 975:switch(f=(s=y).length-10,u=(s.charCodeAt(f)===33?s.substring(0,f):s).substring(y.indexOf(":",7)+1).trim(),_=u.charCodeAt(0)+(u.charCodeAt(7)|0)){case 203:if(111>u.charCodeAt(8))break;case 115:s=s.replace(u,"-webkit-"+u)+";"+s;break;case 207:case 102:s=s.replace(u,"-webkit-"+(102<_?"inline-":"")+"box")+";"+s.replace(u,"-webkit-"+u)+";"+s.replace(u,"-ms-"+u+"box")+";"+s}return s+";";case 938:if(s.charCodeAt(5)===45)switch(s.charCodeAt(6)){case 105:return u=s.replace("-items",""),"-webkit-"+s+"-webkit-box-"+u+"-ms-flex-"+u+s;case 115:return"-webkit-"+s+"-ms-flex-item-"+s.replace(I,"")+s;default:return"-webkit-"+s+"-ms-flex-line-pack"+s.replace("align-content","").replace(I,"")+s}break;case 973:case 989:if(s.charCodeAt(3)!==45||s.charCodeAt(4)===122)break;case 931:case 953:if(ne.test(y)===!0)return(u=y.substring(y.indexOf(":")+1)).charCodeAt(0)===115?o(y.replace("stretch","fill-available"),f,g,C).replace(":fill-available",":stretch"):s.replace(u,"-webkit-"+u)+s.replace(u,"-moz-"+u.replace("fill-",""))+s;break;case 962:if(s="-webkit-"+s+(s.charCodeAt(5)===102?"-ms-"+s:"")+s,g+C===211&&s.charCodeAt(13)===105&&0<s.indexOf("transform",10))return s.substring(0,s.indexOf(";",27)+1).replace(b,"$1-webkit-$2")+s}return s}function i(y,f){var g=y.indexOf(f===1?":":"{"),C=y.substring(0,f!==3?g:10);return g=y.substring(g+1,y.length-1),xe(f!==2?C:C.replace(U,"$1"),g,f)}function a(y,f){var g=o(f,f.charCodeAt(0),f.charCodeAt(1),f.charCodeAt(2));return g!==f+";"?g.replace(R," or ($1)").substring(4):"("+f+")"}function c(y,f,g,C,s,_,u,B,L,N){for(var k=0,X=f,oe;k<se;++k)switch(oe=ce[k].call(m,y,X,g,C,s,_,u,B,L,N)){case void 0:case!1:case!0:case null:break;default:X=oe}if(X!==f)return X}function l(y){switch(y){case void 0:case null:se=ce.length=0;break;default:if(typeof y=="function")ce[se++]=y;else if(typeof y=="object")for(var f=0,g=y.length;f<g;++f)l(y[f]);else ye=!!y|0}return l}function d(y){return y=y.prefix,y!==void 0&&(xe=null,y?typeof y!="function"?Y=1:(Y=2,xe=y):Y=0),d}function m(y,f){var g=y;if(33>g.charCodeAt(0)&&(g=g.trim()),Ae=g,g=[Ae],0<se){var C=c(-1,f,g,g,Q,H,0,0,0,0);C!==void 0&&typeof C=="string"&&(f=C)}var s=t(me,g,f,0,0);return 0<se&&(C=c(-2,s,g,g,Q,H,s.length,0,0,0),C!==void 0&&(s=C)),Ae="",ee=0,H=Q=1,s}var S=/^\0+/g,P=/[\0\r\f]/g,D=/: */g,E=/zoo|gra/,b=/([,: ])(transform)/g,p=/,\r+?/g,A=/([\t\r\n ])*\f?&/g,w=/@(k\w+)\s*(\S*)\s*/,z=/::(place)/g,F=/:(read-only)/g,v=/[svh]\w+-[tblr]{2}/,O=/\(\s*(.*)\s*\)/g,R=/([\s\S]*?);/g,I=/-self|flex-/g,U=/[^]*?(:[rp][el]a[\w-]+)[^]*/,ne=/stretch|:\s*\w+\-(?:conte|avail)/,ue=/([^-])(image-set\()/,H=1,Q=1,ee=0,Y=1,me=[],ce=[],se=0,xe=null,ye=0,Ae="";return m.use=l,m.set=d,e!==void 0&&d(e),m}var Zr={animationIterationCount:1,borderImageOutset:1,borderImageSlice:1,borderImageWidth:1,boxFlex:1,boxFlexGroup:1,boxOrdinalGroup:1,columnCount:1,columns:1,flex:1,flexGrow:1,flexPositive:1,flexShrink:1,flexNegative:1,flexOrder:1,gridRow:1,gridRowEnd:1,gridRowSpan:1,gridRowStart:1,gridColumn:1,gridColumnEnd:1,gridColumnSpan:1,gridColumnStart:1,msGridRow:1,msGridRowSpan:1,msGridColumn:1,msGridColumnSpan:1,fontWeight:1,lineHeight:1,opacity:1,order:1,orphans:1,tabSize:1,widows:1,zIndex:1,zoom:1,WebkitLineClamp:1,fillOpacity:1,floodOpacity:1,stopOpacity:1,strokeDasharray:1,strokeDashoffset:1,strokeMiterlimit:1,strokeOpacity:1,strokeWidth:1};function Jr(e){var t=Object.create(null);return function(r){return t[r]===void 0&&(t[r]=e(r)),t[r]}}var Kr=/^((children|dangerouslySetInnerHTML|key|ref|autoFocus|defaultValue|defaultChecked|innerHTML|suppressContentEditableWarning|suppressHydrationWarning|valueLink|abbr|accept|acceptCharset|accessKey|action|allow|allowUserMedia|allowPaymentRequest|allowFullScreen|allowTransparency|alt|async|autoComplete|autoPlay|capture|cellPadding|cellSpacing|challenge|charSet|checked|cite|classID|className|cols|colSpan|content|contentEditable|contextMenu|controls|controlsList|coords|crossOrigin|data|dateTime|decoding|default|defer|dir|disabled|disablePictureInPicture|download|draggable|encType|enterKeyHint|form|formAction|formEncType|formMethod|formNoValidate|formTarget|frameBorder|headers|height|hidden|high|href|hrefLang|htmlFor|httpEquiv|id|inputMode|integrity|is|keyParams|keyType|kind|label|lang|list|loading|loop|low|marginHeight|marginWidth|max|maxLength|media|mediaGroup|method|min|minLength|multiple|muted|name|nonce|noValidate|open|optimum|pattern|placeholder|playsInline|poster|preload|profile|radioGroup|readOnly|referrerPolicy|rel|required|reversed|role|rows|rowSpan|sandbox|scope|scoped|scrolling|seamless|selected|shape|size|sizes|slot|span|spellCheck|src|srcDoc|srcLang|srcSet|start|step|style|summary|tabIndex|target|title|translate|type|useMap|value|width|wmode|wrap|about|datatype|inlist|prefix|property|resource|typeof|vocab|autoCapitalize|autoCorrect|autoSave|color|incremental|fallback|inert|itemProp|itemScope|itemType|itemID|itemRef|on|option|results|security|unselectable|accentHeight|accumulate|additive|alignmentBaseline|allowReorder|alphabetic|amplitude|arabicForm|ascent|attributeName|attributeType|autoReverse|azimuth|baseFrequency|baselineShift|baseProfile|bbox|begin|bias|by|calcMode|capHeight|clip|clipPathUnits|clipPath|clipRule|colorInterpolation|colorInterpolationFilters|colorProfile|colorRendering|contentScriptType|contentStyleType|cursor|cx|cy|d|decelerate|descent|diffuseConstant|direction|display|divisor|dominantBaseline|dur|dx|dy|edgeMode|elevation|enableBackground|end|exponent|externalResourcesRequired|fill|fillOpacity|fillRule|filter|filterRes|filterUnits|floodColor|floodOpacity|focusable|fontFamily|fontSize|fontSizeAdjust|fontStretch|fontStyle|fontVariant|fontWeight|format|from|fr|fx|fy|g1|g2|glyphName|glyphOrientationHorizontal|glyphOrientationVertical|glyphRef|gradientTransform|gradientUnits|hanging|horizAdvX|horizOriginX|ideographic|imageRendering|in|in2|intercept|k|k1|k2|k3|k4|kernelMatrix|kernelUnitLength|kerning|keyPoints|keySplines|keyTimes|lengthAdjust|letterSpacing|lightingColor|limitingConeAngle|local|markerEnd|markerMid|markerStart|markerHeight|markerUnits|markerWidth|mask|maskContentUnits|maskUnits|mathematical|mode|numOctaves|offset|opacity|operator|order|orient|orientation|origin|overflow|overlinePosition|overlineThickness|panose1|paintOrder|pathLength|patternContentUnits|patternTransform|patternUnits|pointerEvents|points|pointsAtX|pointsAtY|pointsAtZ|preserveAlpha|preserveAspectRatio|primitiveUnits|r|radius|refX|refY|renderingIntent|repeatCount|repeatDur|requiredExtensions|requiredFeatures|restart|result|rotate|rx|ry|scale|seed|shapeRendering|slope|spacing|specularConstant|specularExponent|speed|spreadMethod|startOffset|stdDeviation|stemh|stemv|stitchTiles|stopColor|stopOpacity|strikethroughPosition|strikethroughThickness|string|stroke|strokeDasharray|strokeDashoffset|strokeLinecap|strokeLinejoin|strokeMiterlimit|strokeOpacity|strokeWidth|surfaceScale|systemLanguage|tableValues|targetX|targetY|textAnchor|textDecoration|textRendering|textLength|to|transform|u1|u2|underlinePosition|underlineThickness|unicode|unicodeBidi|unicodeRange|unitsPerEm|vAlphabetic|vHanging|vIdeographic|vMathematical|values|vectorEffect|version|vertAdvY|vertOriginX|vertOriginY|viewBox|viewTarget|visibility|widths|wordSpacing|writingMode|x|xHeight|x1|x2|xChannelSelector|xlinkActuate|xlinkArcrole|xlinkHref|xlinkRole|xlinkShow|xlinkTitle|xlinkType|xmlBase|xmlns|xmlnsXlink|xmlLang|xmlSpace|y|y1|y2|yChannelSelector|z|zoomAndPan|for|class|autofocus)|(([Dd][Aa][Tt][Aa]|[Aa][Rr][Ii][Aa]|x)-.*))$/,Ot=Jr(function(e){return Kr.test(e)||e.charCodeAt(0)===111&&e.charCodeAt(1)===110&&e.charCodeAt(2)<91}),vt=Pe,qr={childContextTypes:!0,contextType:!0,contextTypes:!0,defaultProps:!0,displayName:!0,getDefaultProps:!0,getDerivedStateFromError:!0,getDerivedStateFromProps:!0,mixins:!0,propTypes:!0,type:!0},Qr={name:!0,length:!0,prototype:!0,caller:!0,callee:!0,arguments:!0,arity:!0},en={$$typeof:!0,render:!0,defaultProps:!0,displayName:!0,propTypes:!0},Xt={$$typeof:!0,compare:!0,defaultProps:!0,displayName:!0,propTypes:!0,type:!0},wt={};wt[vt.ForwardRef]=en;wt[vt.Memo]=Xt;function Tt(e){return vt.isMemo(e)?Xt:wt[e.$$typeof]||qr}var tn=Object.defineProperty,rn=Object.getOwnPropertyNames,Ft=Object.getOwnPropertySymbols,nn=Object.getOwnPropertyDescriptor,on=Object.getPrototypeOf,_t=Object.prototype;function Zt(e,t,r){if(typeof t!="string"){if(_t){var n=on(t);n&&n!==_t&&Zt(e,n,r)}var o=rn(t);Ft&&(o=o.concat(Ft(t)));for(var i=Tt(e),a=Tt(t),c=0;c<o.length;++c){var l=o[c];if(!Qr[l]&&!(r&&r[l])&&!(a&&a[l])&&!(i&&i[l])){var d=nn(t,l);try{tn(e,l,d)}catch{}}}}return e}var an=Zt;function de(){return(de=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var r=arguments[t];for(var n in r)Object.prototype.hasOwnProperty.call(r,n)&&(e[n]=r[n])}return e}).apply(this,arguments)}var Lt=function(e,t){for(var r=[e[0]],n=0,o=t.length;n<o;n+=1)r.push(t[n],e[n+1]);return r},st=function(e){return e!==null&&typeof e=="object"&&(e.toString?e.toString():Object.prototype.toString.call(e))==="[object Object]"&&!Pe.typeOf(e)},Me=Object.freeze([]),he=Object.freeze({});function Ee(e){return typeof e=="function"}function jt(e){return e.displayName||e.name||"Component"}function bt(e){return e&&typeof e.styledComponentId=="string"}var be=typeof process<"u"&&{}!==void 0&&({}.REACT_APP_SC_ATTR||{}.SC_ATTR)||"data-styled",St=typeof window<"u"&&"HTMLElement"in window,sn=Boolean(typeof SC_DISABLE_SPEEDY=="boolean"?SC_DISABLE_SPEEDY:typeof process<"u"&&{}!==void 0&&({}.REACT_APP_SC_DISABLE_SPEEDY!==void 0&&{}.REACT_APP_SC_DISABLE_SPEEDY!==""?{}.REACT_APP_SC_DISABLE_SPEEDY!=="false"&&{}.REACT_APP_SC_DISABLE_SPEEDY:{}.SC_DISABLE_SPEEDY!==void 0&&{}.SC_DISABLE_SPEEDY!==""?{}.SC_DISABLE_SPEEDY!=="false"&&{}.SC_DISABLE_SPEEDY:!1));function Re(e){for(var t=arguments.length,r=new Array(t>1?t-1:0),n=1;n<t;n++)r[n-1]=arguments[n];throw new Error("An error occurred. See https://git.io/JUIaE#"+e+" for more information."+(r.length>0?" Args: "+r.join(", "):""))}var cn=function(){function e(r){this.groupSizes=new Uint32Array(512),this.length=512,this.tag=r}var t=e.prototype;return t.indexOfGroup=function(r){for(var n=0,o=0;o<r;o++)n+=this.groupSizes[o];return n},t.insertRules=function(r,n){if(r>=this.groupSizes.length){for(var o=this.groupSizes,i=o.length,a=i;r>=a;)(a<<=1)<0&&Re(16,""+r);this.groupSizes=new Uint32Array(a),this.groupSizes.set(o),this.length=a;for(var c=i;c<a;c++)this.groupSizes[c]=0}for(var l=this.indexOfGroup(r+1),d=0,m=n.length;d<m;d++)this.tag.insertRule(l,n[d])&&(this.groupSizes[r]++,l++)},t.clearGroup=function(r){if(r<this.length){var n=this.groupSizes[r],o=this.indexOfGroup(r),i=o+n;this.groupSizes[r]=0;for(var a=o;a<i;a++)this.tag.deleteRule(o)}},t.getGroup=function(r){var n="";if(r>=this.length||this.groupSizes[r]===0)return n;for(var o=this.groupSizes[r],i=this.indexOfGroup(r),a=i+o,c=i;c<a;c++)n+=this.tag.getRule(c)+`/*!sc*/
-`;return n},e}(),Ne=new Map,De=new Map,ke=1,Le=function(e){if(Ne.has(e))return Ne.get(e);for(;De.has(ke);)ke++;var t=ke++;return Ne.set(e,t),De.set(t,e),t},ln=function(e){return De.get(e)},un=function(e,t){t>=ke&&(ke=t+1),Ne.set(e,t),De.set(t,e)},dn="style["+be+'][data-styled-version="5.3.9"]',pn=new RegExp("^"+be+'\\.g(\\d+)\\[id="([\\w\\d-]+)"\\].*?"([^"]*)'),fn=function(e,t,r){for(var n,o=r.split(","),i=0,a=o.length;i<a;i++)(n=o[i])&&e.registerName(t,n)},hn=function(e,t){for(var r=(t.textContent||"").split(`/*!sc*/
-`),n=[],o=0,i=r.length;o<i;o++){var a=r[o].trim();if(a){var c=a.match(pn);if(c){var l=0|parseInt(c[1],10),d=c[2];l!==0&&(un(d,l),fn(e,d,c[3]),e.getTag().insertRules(l,n)),n.length=0}else n.push(a)}}},mn=function(){return typeof __webpack_nonce__<"u"?__webpack_nonce__:null},Jt=function(e){var t=document.head,r=e||t,n=document.createElement("style"),o=function(c){for(var l=c.childNodes,d=l.length;d>=0;d--){var m=l[d];if(m&&m.nodeType===1&&m.hasAttribute(be))return m}}(r),i=o!==void 0?o.nextSibling:null;n.setAttribute(be,"active"),n.setAttribute("data-styled-version","5.3.9");var a=mn();return a&&n.setAttribute("nonce",a),r.insertBefore(n,i),n},gn=function(){function e(r){var n=this.element=Jt(r);n.appendChild(document.createTextNode("")),this.sheet=function(o){if(o.sheet)return o.sheet;for(var i=document.styleSheets,a=0,c=i.length;a<c;a++){var l=i[a];if(l.ownerNode===o)return l}Re(17)}(n),this.length=0}var t=e.prototype;return t.insertRule=function(r,n){try{return this.sheet.insertRule(n,r),this.length++,!0}catch{return!1}},t.deleteRule=function(r){this.sheet.deleteRule(r),this.length--},t.getRule=function(r){var n=this.sheet.cssRules[r];return n!==void 0&&typeof n.cssText=="string"?n.cssText:""},e}(),yn=function(){function e(r){var n=this.element=Jt(r);this.nodes=n.childNodes,this.length=0}var t=e.prototype;return t.insertRule=function(r,n){if(r<=this.length&&r>=0){var o=document.createTextNode(n),i=this.nodes[r];return this.element.insertBefore(o,i||null),this.length++,!0}return!1},t.deleteRule=function(r){this.element.removeChild(this.nodes[r]),this.length--},t.getRule=function(r){return r<this.length?this.nodes[r].textContent:""},e}(),vn=function(){function e(r){this.rules=[],this.length=0}var t=e.prototype;return t.insertRule=function(r,n){return r<=this.length&&(this.rules.splice(r,0,n),this.length++,!0)},t.deleteRule=function(r){this.rules.splice(r,1),this.length--},t.getRule=function(r){return r<this.length?this.rules[r]:""},e}(),Nt=St,wn={isServer:!St,useCSSOMInjection:!sn},Kt=function(){function e(r,n,o){r===void 0&&(r=he),n===void 0&&(n={}),this.options=de({},wn,{},r),this.gs=n,this.names=new Map(o),this.server=!!r.isServer,!this.server&&St&&Nt&&(Nt=!1,function(i){for(var a=document.querySelectorAll(dn),c=0,l=a.length;c<l;c++){var d=a[c];d&&d.getAttribute(be)!=="active"&&(hn(i,d),d.parentNode&&d.parentNode.removeChild(d))}}(this))}e.registerId=function(r){return Le(r)};var t=e.prototype;return t.reconstructWithOptions=function(r,n){return n===void 0&&(n=!0),new e(de({},this.options,{},r),this.gs,n&&this.names||void 0)},t.allocateGSInstance=function(r){return this.gs[r]=(this.gs[r]||0)+1},t.getTag=function(){return this.tag||(this.tag=(o=(n=this.options).isServer,i=n.useCSSOMInjection,a=n.target,r=o?new vn(a):i?new gn(a):new yn(a),new cn(r)));var r,n,o,i,a},t.hasNameForId=function(r,n){return this.names.has(r)&&this.names.get(r).has(n)},t.registerName=function(r,n){if(Le(r),this.names.has(r))this.names.get(r).add(n);else{var o=new Set;o.add(n),this.names.set(r,o)}},t.insertRules=function(r,n,o){this.registerName(r,n),this.getTag().insertRules(Le(r),o)},t.clearNames=function(r){this.names.has(r)&&this.names.get(r).clear()},t.clearRules=function(r){this.getTag().clearGroup(Le(r)),this.clearNames(r)},t.clearTag=function(){this.tag=void 0},t.toString=function(){return function(r){for(var n=r.getTag(),o=n.length,i="",a=0;a<o;a++){var c=ln(a);if(c!==void 0){var l=r.names.get(c),d=n.getGroup(a);if(l&&d&&l.size){var m=be+".g"+a+'[id="'+c+'"]',S="";l!==void 0&&l.forEach(function(P){P.length>0&&(S+=P+",")}),i+=""+d+m+'{content:"'+S+`"}/*!sc*/
-`}}}return i}(this)},e}(),bn=/(a)(d)/gi,$t=function(e){return String.fromCharCode(e+(e>25?39:97))};function ct(e){var t,r="";for(t=Math.abs(e);t>52;t=t/52|0)r=$t(t%52)+r;return($t(t%52)+r).replace(bn,"$1-$2")}var we=function(e,t){for(var r=t.length;r;)e=33*e^t.charCodeAt(--r);return e},qt=function(e){return we(5381,e)};function Sn(e){for(var t=0;t<e.length;t+=1){var r=e[t];if(Ee(r)&&!bt(r))return!1}return!0}var xn=qt("5.3.9"),An=function(){function e(t,r,n){this.rules=t,this.staticRulesId="",this.isStatic=(n===void 0||n.isStatic)&&Sn(t),this.componentId=r,this.baseHash=we(xn,r),this.baseStyle=n,Kt.registerId(r)}return e.prototype.generateAndInjectStyles=function(t,r,n){var o=this.componentId,i=[];if(this.baseStyle&&i.push(this.baseStyle.generateAndInjectStyles(t,r,n)),this.isStatic&&!n.hash)if(this.staticRulesId&&r.hasNameForId(o,this.staticRulesId))i.push(this.staticRulesId);else{var a=Se(this.rules,t,r,n).join(""),c=ct(we(this.baseHash,a)>>>0);if(!r.hasNameForId(o,c)){var l=n(a,"."+c,void 0,o);r.insertRules(o,c,l)}i.push(c),this.staticRulesId=c}else{for(var d=this.rules.length,m=we(this.baseHash,n.hash),S="",P=0;P<d;P++){var D=this.rules[P];if(typeof D=="string")S+=D;else if(D){var E=Se(D,t,r,n),b=Array.isArray(E)?E.join(""):E;m=we(m,b+P),S+=b}}if(S){var p=ct(m>>>0);if(!r.hasNameForId(o,p)){var A=n(S,"."+p,void 0,o);r.insertRules(o,p,A)}i.push(p)}}return i.join(" ")},e}(),In=/^\s*\/\/.*$/gm,Cn=[":","[",".","#"];function kn(e){var t,r,n,o,i=e===void 0?he:e,a=i.options,c=a===void 0?he:a,l=i.plugins,d=l===void 0?Me:l,m=new Xr(c),S=[],P=function(b){function p(A){if(A)try{b(A+"}")}catch{}}return function(A,w,z,F,v,O,R,I,U,ne){switch(A){case 1:if(U===0&&w.charCodeAt(0)===64)return b(w+";"),"";break;case 2:if(I===0)return w+"/*|*/";break;case 3:switch(I){case 102:case 112:return b(z[0]+w),"";default:return w+(ne===0?"/*|*/":"")}case-2:w.split("/*|*/}").forEach(p)}}}(function(b){S.push(b)}),D=function(b,p,A){return p===0&&Cn.indexOf(A[r.length])!==-1||A.match(o)?b:"."+t};function E(b,p,A,w){w===void 0&&(w="&");var z=b.replace(In,""),F=p&&A?A+" "+p+" { "+z+" }":z;return t=w,r=p,n=new RegExp("\\"+r+"\\b","g"),o=new RegExp("(\\"+r+"\\b){2,}"),m(A||!p?"":p,F)}return m.use([].concat(d,[function(b,p,A){b===2&&A.length&&A[0].lastIndexOf(r)>0&&(A[0]=A[0].replace(n,D))},P,function(b){if(b===-2){var p=S;return S=[],p}}])),E.hash=d.length?d.reduce(function(b,p){return p.name||Re(15),we(b,p.name)},5381).toString():"",E}var Qt=Ue.createContext();Qt.Consumer;var er=Ue.createContext(),Pn=(er.Consumer,new Kt),lt=kn();function En(){return G.useContext(Qt)||Pn}function Rn(){return G.useContext(er)||lt}var On=function(){function e(t,r){var n=this;this.inject=function(o,i){i===void 0&&(i=lt);var a=n.name+i.hash;o.hasNameForId(n.id,a)||o.insertRules(n.id,a,i(n.rules,a,"@keyframes"))},this.toString=function(){return Re(12,String(n.name))},this.name=t,this.id="sc-keyframes-"+t,this.rules=r}return e.prototype.getName=function(t){return t===void 0&&(t=lt),this.name+t.hash},e}(),Tn=/([A-Z])/,Fn=/([A-Z])/g,_n=/^ms-/,Ln=function(e){return"-"+e.toLowerCase()};function Mt(e){return Tn.test(e)?e.replace(Fn,Ln).replace(_n,"-ms-"):e}var Dt=function(e){return e==null||e===!1||e===""};function Se(e,t,r,n){if(Array.isArray(e)){for(var o,i=[],a=0,c=e.length;a<c;a+=1)(o=Se(e[a],t,r,n))!==""&&(Array.isArray(o)?i.push.apply(i,o):i.push(o));return i}if(Dt(e))return"";if(bt(e))return"."+e.styledComponentId;if(Ee(e)){if(typeof(d=e)!="function"||d.prototype&&d.prototype.isReactComponent||!t)return e;var l=e(t);return Se(l,t,r,n)}var d;return e instanceof On?r?(e.inject(r,n),e.getName(n)):e:st(e)?function m(S,P){var D,E,b=[];for(var p in S)S.hasOwnProperty(p)&&!Dt(S[p])&&(Array.isArray(S[p])&&S[p].isCss||Ee(S[p])?b.push(Mt(p)+":",S[p],";"):st(S[p])?b.push.apply(b,m(S[p],p)):b.push(Mt(p)+": "+(D=p,(E=S[p])==null||typeof E=="boolean"||E===""?"":typeof E!="number"||E===0||D in Zr?String(E).trim():E+"px")+";"));return P?[P+" {"].concat(b,["}"]):b}(e):e.toString()}var Ut=function(e){return Array.isArray(e)&&(e.isCss=!0),e};function jn(e){for(var t=arguments.length,r=new Array(t>1?t-1:0),n=1;n<t;n++)r[n-1]=arguments[n];return Ee(e)||st(e)?Ut(Se(Lt(Me,[e].concat(r)))):r.length===0&&e.length===1&&typeof e[0]=="string"?e:Ut(Se(Lt(e,r)))}var Nn=function(e,t,r){return r===void 0&&(r=he),e.theme!==r.theme&&e.theme||t||r.theme},$n=/[!"#$%&'()*+,./:;<=>?@[\\\]^`{|}~-]+/g,Mn=/(^-|-$)/g;function rt(e){return e.replace($n,"-").replace(Mn,"")}var Dn=function(e){return ct(qt(e)>>>0)};function je(e){return typeof e=="string"&&!0}var ut=function(e){return typeof e=="function"||typeof e=="object"&&e!==null&&!Array.isArray(e)},Un=function(e){return e!=="__proto__"&&e!=="constructor"&&e!=="prototype"};function zn(e,t,r){var n=e[r];ut(t)&&ut(n)?tr(n,t):e[r]=t}function tr(e){for(var t=arguments.length,r=new Array(t>1?t-1:0),n=1;n<t;n++)r[n-1]=arguments[n];for(var o=0,i=r;o<i.length;o++){var a=i[o];if(ut(a))for(var c in a)Un(c)&&zn(e,a[c],c)}return e}var rr=Ue.createContext();rr.Consumer;var nt={};function nr(e,t,r){var n=bt(e),o=!je(e),i=t.attrs,a=i===void 0?Me:i,c=t.componentId,l=c===void 0?function(w,z){var F=typeof w!="string"?"sc":rt(w);nt[F]=(nt[F]||0)+1;var v=F+"-"+Dn("5.3.9"+F+nt[F]);return z?z+"-"+v:v}(t.displayName,t.parentComponentId):c,d=t.displayName,m=d===void 0?function(w){return je(w)?"styled."+w:"Styled("+jt(w)+")"}(e):d,S=t.displayName&&t.componentId?rt(t.displayName)+"-"+t.componentId:t.componentId||l,P=n&&e.attrs?Array.prototype.concat(e.attrs,a).filter(Boolean):a,D=t.shouldForwardProp;n&&e.shouldForwardProp&&(D=t.shouldForwardProp?function(w,z,F){return e.shouldForwardProp(w,z,F)&&t.shouldForwardProp(w,z,F)}:e.shouldForwardProp);var E,b=new An(r,S,n?e.componentStyle:void 0),p=b.isStatic&&a.length===0,A=function(w,z){return function(F,v,O,R){var I=F.attrs,U=F.componentStyle,ne=F.defaultProps,ue=F.foldedComponentIds,H=F.shouldForwardProp,Q=F.styledComponentId,ee=F.target,Y=function(C,s,_){C===void 0&&(C=he);var u=de({},s,{theme:C}),B={};return _.forEach(function(L){var N,k,X,oe=L;for(N in Ee(oe)&&(oe=oe(u)),oe)u[N]=B[N]=N==="className"?(k=B[N],X=oe[N],k&&X?k+" "+X:k||X):oe[N]}),[u,B]}(Nn(v,G.useContext(rr),ne)||he,v,I),me=Y[0],ce=Y[1],se=function(C,s,_,u){var B=En(),L=Rn(),N=s?C.generateAndInjectStyles(he,B,L):C.generateAndInjectStyles(_,B,L);return N}(U,R,me),xe=O,ye=ce.$as||v.$as||ce.as||v.as||ee,Ae=je(ye),y=ce!==v?de({},v,{},ce):v,f={};for(var g in y)g[0]!=="$"&&g!=="as"&&(g==="forwardedAs"?f.as=y[g]:(H?H(g,Ot,ye):!Ae||Ot(g))&&(f[g]=y[g]));return v.style&&ce.style!==v.style&&(f.style=de({},v.style,{},ce.style)),f.className=Array.prototype.concat(ue,Q,se!==Q?se:null,v.className,ce.className).filter(Boolean).join(" "),f.ref=xe,G.createElement(ye,f)}(E,w,z,p)};return A.displayName=m,(E=Ue.forwardRef(A)).attrs=P,E.componentStyle=b,E.displayName=m,E.shouldForwardProp=D,E.foldedComponentIds=n?Array.prototype.concat(e.foldedComponentIds,e.styledComponentId):Me,E.styledComponentId=S,E.target=n?e.target:e,E.withComponent=function(w){var z=t.componentId,F=function(O,R){if(O==null)return{};var I,U,ne={},ue=Object.keys(O);for(U=0;U<ue.length;U++)I=ue[U],R.indexOf(I)>=0||(ne[I]=O[I]);return ne}(t,["componentId"]),v=z&&z+"-"+(je(w)?w:rt(jt(w)));return nr(w,de({},F,{attrs:P,componentId:v}),r)},Object.defineProperty(E,"defaultProps",{get:function(){return this._foldedDefaultProps},set:function(w){this._foldedDefaultProps=n?tr({},e.defaultProps,w):w}}),Object.defineProperty(E,"toString",{value:function(){return"."+E.styledComponentId}}),o&&an(E,e,{attrs:!0,componentStyle:!0,displayName:!0,foldedComponentIds:!0,shouldForwardProp:!0,styledComponentId:!0,target:!0,withComponent:!0}),E}var dt=function(e){return function t(r,n,o){if(o===void 0&&(o=he),!Pe.isValidElementType(n))return Re(1,String(n));var i=function(){return r(n,o,jn.apply(void 0,arguments))};return i.withConfig=function(a){return t(r,n,de({},o,{},a))},i.attrs=function(a){return t(r,n,de({},o,{attrs:Array.prototype.concat(o.attrs,a).filter(Boolean)}))},i}(nr,e)};["a","abbr","address","area","article","aside","audio","b","base","bdi","bdo","big","blockquote","body","br","button","canvas","caption","cite","code","col","colgroup","data","datalist","dd","del","details","dfn","dialog","div","dl","dt","em","embed","fieldset","figcaption","figure","footer","form","h1","h2","h3","h4","h5","h6","head","header","hgroup","hr","html","i","iframe","img","input","ins","kbd","keygen","label","legend","li","link","main","map","mark","marquee","menu","menuitem","meta","meter","nav","noscript","object","ol","optgroup","option","output","p","param","picture","pre","progress","q","rp","rt","ruby","s","samp","script","section","select","small","source","span","strong","style","sub","summary","sup","table","tbody","td","textarea","tfoot","th","thead","time","title","tr","track","u","ul","var","video","wbr","circle","clipPath","defs","ellipse","foreignObject","g","image","line","linearGradient","marker","mask","path","pattern","polygon","polyline","radialGradient","rect","stop","svg","text","textPath","tspan"].forEach(function(e){dt[e]=dt(e)});const M=dt,Gn=M.div`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 255px;
-    overflow: hidden;
-`,Bn=M.div`
-    position: relative;
-    text-decoration: none;
-    display: block;
-`,Vn=M.div`
-    width: 175px;
-    min-height: 158px;
-    border-radius: 4px;
-    padding: 5px 5px 5px 5px;
-    margin: 0rem 1.0rem .5rem 2.0rem;
-    opacity: 1;
-    background-color: 1;
-    transition: box-shadow 100ms linear 0s;
-`,zt=M.div`
-    display: flex;
-`,Hn=M.div`
-    display: flex;
-    flex-direction: column;
-    margin: 0px 5px 0px 0px;
-`,Wn=M.span`
-    font-family: Inter;
-    font-size: 12px;
-    font-weight: 600;
-    line-height: 20px;
-    letter-spacing: -0.006em;
-    color: rgb(42, 51, 61)
-`,Yn=M.span`
-    font-family: Inter;
-    font-size: 10px;
-    line-height: 16px;
-    color: rgb(91, 102, 113)
-`,Xn=M.div`
-    custor: pointer;
-    height: 115px;
-    width: 115px;
-    border-radius: 4px;
-    background-size: cover;
-    background-color: white;
-    background-position: center center;
-    overflow: hidden;
-`,Zn=M.div`
-    transition: opacity 100ms linear 0s;
-    overflow: hidden;
-    background-color: rgb(226, 226, 226);
-    display: flex;
-    -webkit-box-align: center;
-    align-items: center;
-    -webkit-box-pack: center;
-    justify-content: center;
-    height: 100%;
-    width: 100%;
-`,Jn=M.div`
-    display: flex;
-    flex-direction: column;
-    margin: 0rem .1rem 0rem .4rem;
-    padding: 0px .2em 0px .5em;
-    border: 1.5px solid rgb(233, 234, 236);
-    border-radius: 4px;
-    align-items: center;
-    justify-content: center;
-`;M.div`
-    height: 80px;
-    width: 80px;
-    display: flex;
-    justify-content: space-evenly;
-    -webkit-box-align: center;
-    align-items: center;
-    flex-direction: column;
-    margin-bottom: 10px;
-`;const Kn=M.div`
-    margin-top: 6px;
-`,qn=M.div`
-    font-family: Inter;
-    font-size: 12px;
-    font-weight: 550;
-    color: rgb(42, 51, 61);
-    line-height: 24px;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-    text-align: left;
-`,Qn=M.div`
-    position: relative;
-    width: fit-content;
-    text-align: left;
-`,eo=M.div`
-    position: relative;
-    font-family: Inter;
-    font-size: 10px;
-    line-height: 16px;
-    color: rgb(115, 127, 140);
-    margin-top: -2px;
-    min-height: 16px;
-    width: fit-content;
-    overflow: visible;
-    text-align: left;
-`,to=M.div`
-    display: flex;
-    flex-direction: column;
-    flex-row: row wrap;
-    overflow: visible;
-    align-items: center;
-`,ro=M.div`
-    position: relative;
-    text-decoration: none;
-    display: block;
-`,no=M.div`
-    width: 175px;
-    border-radius: 4px;
-    padding: 5px 5px 5px 5px;
-    margin: 0rem 1.0rem .5rem 2.0rem;
-    opacity: 1;
-    background-color: 1;
-    transition: box-shadow 100ms linear 0s;
-    overflow-y: auto;
-`,Gt=M.div`
-    display: flex;
-`,oo=M.div`
-    display: flex;
-    flex-direction: column;
-    margin: 0px 5px 0px 0px;
-`,io=M.div`
-    custor: pointer;
-    height: 56px;
-    width: 56px;
-    border-radius: 4px;
-    background-size: cover;
-    background-color: white;
-    background-position: center center;
-    overflow: hidden;
-`,ao=M.div`
-    transition: opacity 100ms linear 0s;
-    overflow: hidden;
-    background-color: rgb(226, 226, 226);
-    display: flex;
-    -webkit-box-align: center;
-    align-items: center;
-    -webkit-box-pack: center;
-    justify-content: center;
-    height: 100%;
-    width: 100%;
-`,so=M.div`
-    display: flex;
-    flex-direction: column;
-    margin: 0rem .1rem 0rem .4rem;
-    padding: 0px .2em 0px .5em;
-    border-radius: 4px;
-    align-items: center;
-    justify-content: center;
-`;M.div`
-    height: 80px;
-    width: 80px;
-    display: flex;
-    justify-content: space-evenly;
-    -webkit-box-align: center;
-    align-items: center;
-    flex-direction: column;
-    margin-bottom: 10px;
-`;M.div`
-    margin-top: 6px;
-`;const co=M.div`
-    font-family: Inter;
-    font-size: 12px;
-    font-weight: 550;
-    color: rgb(42, 51, 61);
-    line-height: 24px;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-    text-align: left;
-`,lo=M.div`
-    position: relative;
-    width: fit-content;
-    text-align: left;
-`,uo=M.div`
-    position: relative;
-    font-family: Inter;
-    font-size: 10px;
-    line-height: 16px;
-    color: rgb(115, 127, 140);
-    margin-top: -2px;
-    min-height: 16px;
-    width: fit-content;
-    overflow: visible;
-    text-align: left;
-`;function po(e){const{proposal:t,proposalSelectionHandler:r,isSelected:n}=e,o=ve.FormatThumbnailUrl(t.projectId,t.urn);return h(pe,{children:h(ro,{onClick:()=>{r(n?"":t.id)},children:h(no,{id:t.projectId,className:n?"proposal selected":"proposal",children:ie(Gt,{className:"infos",children:[h(io,{children:h(ao,{children:h("img",{alt:"Site Preview",src:o,width:"56",height:"56"})})}),h(so,{children:h(Gt,{children:ie(oo,{children:[h(co,{children:t.name}),h(lo,{children:h(uo,{children:t.creationDate})})]})})})]})})})})}function fo(e){const{proposals:t,proposalSelectionHandler:r,selectedProposalId:n}=e;return t?h(pe,{children:h("div",{"data-checkly":"ProposalList",children:h(to,{children:t==null?void 0:t.map(o=>h(po,{proposal:o,proposalSelectionHandler:r,isSelected:o.proposalId===n},o.proposalId))})})}):h(pe,{children:h("div",{"data-checkly":"ProposalList"})})}function ho(e){const{project:t,projectSelectionHandler:r,proposalSelectionHandler:n,isSelected:o,selectedProposalId:i}=e,a=ve.FormatThumbnailUrl(t.projectId,t.urn);return ie(pe,{children:[h(Bn,{onClick:()=>{r(o?"":t.id)},children:ie(Vn,{id:t.projectId,className:o?"project selected":"project",children:[ie(zt,{className:"infos",children:[h(Xn,{children:h(Zn,{children:h("img",{alt:"Site Preview",src:a,width:"115",height:"115"})})}),h(Jn,{children:h(zt,{children:ie(Hn,{children:[h(Wn,{children:t.proposalCount}),h(Yn,{children:"Proposals"})]})})})]}),ie(Kn,{children:[h(qn,{children:t.name}),h(Qn,{children:h(eo,{children:t.creationTime})})]})]})}),h("div",{id:t.proposalsListContainer,className:o?"proposal-visible":"proposal-hidden",children:h(fo,{proposals:t.proposals,proposalSelectionHandler:n,selectedProposalId:i})})]})}function mo(e){const{projects:t,projectSelectionHandler:r,proposalSelectionHandler:n,selectedProjectId:o,selectedProposalId:i}=e;if(!t)return h(pe,{children:h("div",{"data-checkly":"ProjectList"})});function a(l){document.querySelectorAll(".project.selected").forEach(S=>{S.classList.remove("selected")}),document.querySelectorAll(".proposal-visible").forEach(S=>{S.classList.remove("proposal-visible"),S.classList.add("proposal-hidden")}),c(""),r(l)}function c(l){document.querySelectorAll(".proposal.selected").forEach(m=>{m.classList.remove("selected")}),n(l)}return h(pe,{children:h("div",{id:"projectlist-container","data-checkly":"ProjectList",children:h(Gn,{children:t==null?void 0:t.map(l=>h(ho,{project:l,projectSelectionHandler:a,proposalSelectionHandler:c,isSelected:l.projectId===o,selectedProposalId:i},l.projectId))})})})}class go extends ht{constructor(){super(...arguments);te(this,"projectId");te(this,"newestProposalId");te(this,"creationTime");te(this,"proposalCount");te(this,"proposals");te(this,"proposalsListContainer")}Fill(r,n,o,i){super.Fill(r,n,o,i,""),this.projectId=this.id,this.newestProposalId="",this.creationTime="",this.proposalCount=0}}class yo extends ht{constructor(){super(...arguments);te(this,"proposalId");te(this,"projectId");te(this,"creationDate")}Fill(r,n,o,i,a){super.Fill(r,n,o,i,a),this.proposalId=this.id,this.creationDate=""}}function vo(){}function wo(){function e(v,O){v.then(O).catch(R=>console.log(R))}function t(v){try{let O=v!==null&&v.length>0;if(document.getElementById("message").style.display=O?"block":"none",O){P("info"),m("");let R=v.map(I=>{var U=new ht;return U.Fill(I.id,I.name,I.version,I.metadata,I.urn),U});E(R),R.length>0&&o(R[0].id)}}catch{const R="Unable to read workspaces";throw P("error"),m(R),new Error(R)}}async function r(v){try{let O=v.filter(I=>I.metadata!==null&&!I.metadata.isDraft&&I.version!==1).map(async I=>{var U=new go;U.Fill(I.id,I.name,I.version,I.metadata);let ue=new Date().getTime()-I.created,H=Math.floor(ue/(1e3*24*60*60));return U.creationTime=H==0?"Today":`${H} days ago`,await ve.countProposals(U.projectId).then(Q=>{U.proposalCount=Q}),U}),R=await Promise.all(O);w(R),P("info"),m("")}catch{const R="Unable to read projects from selected workspace";throw P("error"),m(R),new Error(R)}}async function n(v){try{P("info"),m("Fetching proposals...");let O;await ve.getProposals(v).then(R=>{O=R.map(I=>{var U=new yo;const ne=I.urn.split(":");let ue=ne[ne.length-2],H=ne[ne.length-1],Q=I.properties.name;U.Fill(ue,Q,H,I.metadata,I.urn);let ee=new Date(I.metadata.createdAt),Y=ee.getHours(),me=ee.getMinutes();return ee.getDate()===new Date().getDate()?U.creationDate=`Today ${Y}:${me} ${Y<12?"AM":"PM"}`:U.creationDate=`${ee.toLocaleString()}`,U.projectId=v,U}),O.sort(function(I,U){return new Date(I.metadata.createdAt).getTime()-new Date(U.metadata.createdAt).getTime()})}),A.forEach(R=>{R.projectId===v&&(R.proposals=O,O.length>0&&(R.newestProposalId=R.proposals[0].id,R.urn=R.proposals[0].urn),R.proposalsListContainer=`project-${R.projectId}-proposals-list`)}),w(A),P("info"),m("")}catch{const R="Unable to read proposals from projects";throw P("error"),m(R),new Error(R)}}function o(v){P("info"),m("Fetching projects..."),e(ve.getProjects(v),r.bind(this))}function i(){let v=document.getElementById("workspace-select");v!==null&&o(v.value)}async function a(v){F(v);const O=await FormIt.Model.IsModified(),R=b!==""&&v!=="";let I=document.getElementById("sync-btn");I!==null&&(I.disabled=!O||!R,(I.onclick===vo||!I.onclick)&&(I.onclick=c))}function c(){let v=document.getElementById("projectlist-container");v.classList.add("disabled");let O=document.getElementById("message");O.className="info",O.textContent="Sending datas to Forma...",ft.save({projectId:b,proposalId:z},R=>{P(R?"success":"error"),m(R?"Datas have been synchronized successfully on Forma":"Synchronization failed"),v.classList.remove("disabled")})}function l(v){A.forEach(O=>{O.projectId===v&&(O.proposals||n(v))}),p(v)}const[d,m]=G.useState(""),[S,P]=G.useState("info"),[D,E]=G.useState(),[b,p]=G.useState(""),[A,w]=G.useState(),[z,F]=G.useState("");return G.useEffect(()=>{m("Fetching worskpaces..."),e(ve.getWorkspaces(),t.bind(this))},[]),h("div",{id:"FormaControls",className:"col-md-12",children:ie("div",{id:"plugin-container",className:"plugin",children:[h("h3",{id:"identifier",children:"Welcome to Formit-Forma plugin"}),h("select",{id:"workspace-select",className:"fetchSelect",onChange:i.bind(this),defaultValue:"",hidden:!0,children:D==null?void 0:D.map(({id:v,name:O})=>h("option",{value:v,children:O},v))}),h(mo,{projects:A,projectSelectionHandler:l,proposalSelectionHandler:a,selectedProjectId:b,selectedProposalId:z}),ie("div",{id:"action",children:[h("button",{className:"st",id:"sync-btn",disabled:!0,children:"Sync"}),h("label",{id:"message",className:S,children:d})]})]})})}function bo(){let e=window.location.href.indexOf("loggedIn=1")>-1,t=ft.getCookie("ajs_user_id"),r=null;return e=e||t!==null,e?r=h(wo,{}):r=ie("div",{id:"LoginControls",className:"",children:[h("h4",{children:"Start plugin to select a project"}),ie("button",{id:"LoginButton",className:"button is-link",onClick:Dr.login,children:[h("span",{children:"Start plugin"}),h("i",{className:"fab fa-github fa-lg"})]})]}),{loggedIn:e,node:r}}function So(){const[e,t]=G.useState(!1),[r,n]=G.useState(h(pe,{}));return G.useEffect(()=>{let o=bo();o!==null&&(t(o.loggedIn),n(o.node))},[]),h("div",{id:"PluginWrapper",children:ie("div",{id:"PluginContainer",children:[h("h1",{className:"title",children:"Formit-Forma"}),h("h3",{className:"title",children:"Send data between FormIt and Forma"}),h("div",{id:"AppControls",children:h("div",{className:"container mt-3",children:h("div",{id:"app",children:r})})}),ie("div",{id:"NoAccess",hidden:!0,children:[h("div",{children:h("div",{className:"alertIcon"})}),ie("div",{children:["It looks like your web browser is in private or incognito mode. Formit-Forma plugin needs to save data to the local storage, so it can only be used in standard browsing mode.",h("br",{}),h("br",{}),"To use Formit-Forma Plugin, please switch to standard browsing mode."]})]})]})})}const xo=ar.createRoot(document.getElementById("root"));xo.render(h(pe,{children:h("div",{id:"MainControls",children:h(Pr,{children:h(So,{})})})}));
