+++ v0_1/index.html
@@ -11,7 +11,7 @@
     <script type="text/javascript">
       FormItInterface.Initialize();
     </script>
-    <script type="module" crossorigin src="./assets/main-75dafeac.js"></script>
+    <script type="module" crossorigin src="./assets/main-2cd36978.js"></script>
     <link rel="modulepreload" crossorigin href="./assets/index-340c65dc.js">
     <link rel="stylesheet" href="./assets/index-70e0b2f7.css">
     <link rel="stylesheet" href="./assets/index-9e4fa14f.css">
